Moved search specific info into handler files
Realised that didn't have to keep the one page open for the cookie so
have rearranged the code to allow the core to be completely independent
of any of the handlers. Had to update the handlers as well as they are
now completely responsible for making their own connections and scraping
the information that they want. Handler signature has changed to just
function (isbn, callback)